+++ modules/clients/src/test/java/org/apache/ignite/internal/client/ClientStopNodeTask.java
@@ -77,7 +77,7 @@ public class ClientStopNodeTask extends ComputeTaskSplitAdapter<String, Integer>
 
         for (ComputeJobResult res : results)
             if (!res.isCancelled())
-                stoppedCnt += (Integer)res.getData();
+                stoppedCnt += res.getData();
 
         return stoppedCnt;
     }
