+++ modules/core/src/main/java/org/apache/ignite/internal/jdbc/thin/JdbcThinResultSet.java
@@ -389,7 +389,7 @@ public class JdbcThinResultSet implements ResultSet {
         if (val instanceof Number)
             return ((Number)val).longValue();
         else if (cls == Boolean.class)
-            return (long)((Boolean)val ? 1 : 0);
+            return (Boolean)val ? 1 : 0;
         else if (cls == String.class || cls == Character.class) {
             try {
                 return Long.parseLong(val.toString());
@@ -439,7 +439,7 @@ public class JdbcThinResultSet implements ResultSet {
         if (val instanceof Number)
             return ((Number)val).doubleValue();
         else if (cls == Boolean.class)
-            return (double)((Boolean)val ? 1 : 0);
+            return (Boolean)val ? 1 : 0;
         else if (cls == String.class || cls == Character.class) {
             try {
                 return Double.parseDouble(val.toString());
