+++ modules/core/src/test/java/org/apache/ignite/p2p/GridP2PRecursionTaskSelfTest.java
@@ -167,7 +167,7 @@ public class GridP2PRecursionTaskSelfTest extends GridCommonAbstractTest {
             long retVal = 0;
 
             for (ComputeJobResult res : results)
-                retVal += (Long)res.getData();
+                retVal += res.getData();
 
             return retVal;
         }
